# site/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Run402 API",
    "version": "2.0.0",
    "description": "Full-stack infrastructure for AI agents. Postgres, REST API, auth, file storage, static site hosting, serverless functions. Pay-per-tier via x402 USDC on Base — subscribe once, then use wallet signatures for all operations within your tier.",
    "contact": {
      "name": "Run402",
      "url": "https://run402.com"
    },
...
  },
  "servers": [
    {
      "url": "https://api.run402.com",
      "description": "Production"
    }
  ],
...
  "paths": {
...
    "/projects/v1/admin/{id}/sql": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Run SQL migration",
        "operationId": "runSql",
        "description": "Executes SQL against the project database. Dangerous statements (CREATE EXTENSION, ALTER SYSTEM, etc.) are blocked.",
        "security": [
          {
            "serviceKey": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "sql"
                ],
                "properties": {
                  "sql": {
                    "type": "string",
                    "description": "SQL statement(s) to execute"
                  }
                }
              }
            },
            "text/plain": {
              "schema": {
                "type": "string",
                "description": "Raw SQL text"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SQL executed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    },
                    "schema": {
                      "type": "string"
                    },
                    "rows": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "rowCount": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/projects/v1/admin/{id}/rls": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Apply RLS policy template",
        "operationId": "applyRls",
        "description": "Applies a row-level security template to one or more tables.",
        "security": [
          {
            "serviceKey": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "template",
                  "tables"
                ],
                "properties": {
                  "template": {
                    "type": "string",
                    "enum": [
                      "user_owns_rows",
                      "public_read",
                      "public_read_write"
                    ],
                    "description": "RLS template to apply"
                  },
                  "tables": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Table names to apply the policy to"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "RLS applied",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    },
                    "template": {
                      "type": "string"
                    },
                    "tables": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/projects/v1/admin/{id}/usage": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get project usage",
        "operationId": "getUsage",
        "security": [
          {
            "serviceKey": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usage stats",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "project_id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "tier": {
                      "type": "string"
                    },
                    "api_calls": {
                      "type": "integer"
                    },
                    "api_calls_limit": {
                      "type": "integer"
                    },
                    "storage_bytes": {
                      "type": "integer"
                    },
                    "storage_limit_bytes": {
                      "type": "integer"
                    },
                    "lease_expires_at": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/projects/v1/admin/{id}/schema": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Introspect project schema",
        "operationId": "getSchema",
        "description": "Returns full schema introspection: tables, columns, constraints, RLS policies.",
        "security": [
          {
            "serviceKey": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Schema info",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "schema": {
                      "type": "string"
                    },
                    "tables": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "name": {
                            "type": "string"
                          },
                          "columns": {
                            "type": "array",
                            "items": {
                              "type": "object"
                            }
                          },
                          "constraints": {
                            "type": "array",
                            "items": {
                              "type": "object"
                            }
                          },
                          "rls_enabled": {
                            "type": "boolean"
                          },
                          "policies": {
                            "type": "array",
                            "items": {
                              "type": "object"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apikey": {
        "type": "apiKey",
        "in": "header",
        "name": "apikey",
        "description": "Project anon_key (JWT). Returned when you provision a project."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "User access_token from /auth/v1/token. Pass as Authorization: Bearer <token>."
      },
      "serviceKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Project service_key (JWT with service_role). Returned when you provision a project. Pass as Authorization: Bearer <service_key>."
      },
      "adminKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-key",
        "description": "Server admin key. Required for admin operations."
      },
      "x402": {
        "type": "apiKey",
        "in": "header",
        "name": "X-PAYMENT",
        "description": "x402 payment header. The client library (e.g. @anthropic-ai/sdk x402 facilitator) handles this automatically. Payment is in USDC on Base (mainnet eip155:8453 or testnet eip155:84532)."
      },
      "walletAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Run402-Wallet",
        "description": "EIP-4361 wallet authentication. Requires three headers: X-Run402-Wallet (wallet address), X-Run402-Signature (signed message), X-Run402-Timestamp (unix timestamp). The wallet must have an active tier subscription."
      }
    },
...
    "responses": {
...
      "Unauthorized": {
        "description": "Missing or invalid authentication. For wallet-auth endpoints, provide X-Run402-Wallet, X-Run402-Signature, and X-Run402-Timestamp headers. For project endpoints, provide apikey header (anon_key) or Authorization: Bearer (service_key/access_token).",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  }
}